move tasks to package.json (also rename bin/*)
