+++ package.json
@@ -8,7 +8,14 @@
   "scripts": {
     "eslint": "eslint src/",
     "mocha": "mocha",
-    "test": "npm run eslint && npm run mocha"
+    "build": "make lib/parse-hanson-string.js",
+    "test": "npm run eslint && npm run mocha",
+    "cover": "istanbul cover ./node_modules/.bin/_mocha",
+    "codecov.io": "istanbul cover ./node_modules/.bin/_mocha -- -R spec && ./node_modules/codecov.io/bin/codecov.io.js < ./coverage/coverage.json",
+    "compile": "babel-node ./bin/compile.js",
+    "evaluate": "babel-node ./bin/evaluate.js",
+    "parse-hanson": "babel-node ./bin/parse-hanson.js",
+    "run": "babel-node ./bin/run.js"
   },
   "dependencies": {
     "archy": "1.0.0",
